Clean up a couple warnings
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@3911 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/app/server/WinBorder.cpp
@@ -302,7 +302,7 @@ printf("ClickMove: Drag\n");
 			BRegion reg2(oldmoveframe);
 			MoveBy(dx,dy);
 			_decorator->MoveBy(BPoint(dx, dy));
-			reg->OffsetBy(dx, dy);
+			reg->OffsetBy((int32)dx, (int32)dy);
 			reg2.Exclude(reg);
 			_parent->RebuildRegions();
 			_parent->RequestDraw();
